use grad student service
(cherry picked from commit 1671dea2d6c96b49727531550087bde898749d13)
